requirements-core.txt
# Core dependencies for AuMeet_package
# AI and ML libraries
openai>=1.0.0
ollama>=0.1.0
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0
hdbscan>=0.8.0
umap-learn>=0.5.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Text processing
zhconv-rs>=0.3.0
jieba>=0.42.1

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# CLI and progress
tqdm>=4.65.0
rich>=13.0.0
click>=8.0.0

# HTTP and async
httpx>=0.24.0
aiohttp>=3.8.0

# Utilities
pathlib2>=2.3.0
requests>=2.28.0